refactor: modify class path & ktlintCheck
